Change CCL input precision from int8 to int64 to align with compiler

+++ QEfficient/transformers/models/mistral3/modeling_mistral3.py
@@ -402,7 +402,7 @@ class QEffMistral3ForConditionalGeneration(Mistral3ForConditionalGeneration):
                 lang_inputs["past_key_values"][i].append(torch.zeros(kv_cache_shape, dtype=self.config.torch_dtype))
 
         if comp_ctx_lengths is not None:
-            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.int8)
+            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.int64)
         if continuous_batching:
             lang_inputs["batch_index"] = torch.arange(bs).view(bs, 1)
